Итог

+++ src/main/java/ru/technoserv/repository/EmployeeRepositoryImpl.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.stereotype.Component;
-import ru.technoserv.controller.Employee;
+import ru.technoserv.dao.Employee;
 import ru.technoserv.services.EmployeeService;
 
 @Component
@@ -17,14 +17,14 @@ public class EmployeeRepositoryImpl implements EmployeeRepository{
     @Override
     @Cacheable("employees")
     public Employee getEmployeeById(int id){
-        Employee e =employeeService.getEmployee(id);
+        Employee e = employeeService.getEmployee(id);
         simulateSlowService();
         return e;
     }
 
     private void simulateSlowService() {
         try {
-            long time = (long) (5000L);
+            long time = 5000L;
             System.out.println("Sleep");
             Thread.sleep(time);
         } catch (InterruptedException e) {
